Move Entities and Models to Domain namespace
Also preformed code cleanup after running static analysis tests.
Updated the privacy statement.
