+++ src/Surfnet/ServiceProviderDashboard/Domain/Model/Attribute.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2017 SURFnet bv
+ * Copyright 2017 SURFnet B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,8 @@
  * limitations under the License.
  */
 
-namespace Surfnet\ServiceProviderDashboard\Model;
+namespace Surfnet\ServiceProviderDashboard\Domain\Model;
 
-/**
- * Class Attribute
- */
 class Attribute
 {
     /**
